Move non-core-migration tests
