+++ runtime/src/bank/tests.rs
@@ -84,7 +84,8 @@ use {
     solana_rent::Rent,
     solana_reward_info::RewardType,
     solana_sdk_ids::{
-        bpf_loader, bpf_loader_upgradeable, incinerator, native_loader, secp256k1_program,
+        bpf_loader, bpf_loader_upgradeable, ed25519_program, incinerator, native_loader,
+        secp256k1_program,
     },
     solana_sha256_hasher::hash,
     solana_signature::Signature,
@@ -12370,4 +12371,59 @@ fn test_should_use_vote_keyed_leader_schedule() {
             assert_eq!(test_bank.should_use_vote_keyed_leader_schedule(epoch), None);
         }
     }
+}
+
+#[test]
+fn test_apply_builtin_program_feature_transitions_for_new_epoch() {
+    let (genesis_config, _mint_keypair) = create_genesis_config(100_000);
+
+    let mut bank = Bank::new_for_tests(&genesis_config);
+    bank.feature_set = Arc::new(FeatureSet::all_enabled());
+    bank.finish_init(&genesis_config, None, false);
+
+    // Overwrite precompile accounts to simulate a cluster which already added precompiles.
+    for precompile in get_precompiles() {
+        bank.store_account(&precompile.program_id, &AccountSharedData::default());
+        // Simulate cluster which added ed25519 precompile with a system program owner
+        if precompile.program_id == ed25519_program::id() {
+            bank.add_precompiled_account_with_owner(
+                &precompile.program_id,
+                solana_system_interface::program::id(),
+            );
+        } else {
+            bank.add_precompiled_account(&precompile.program_id);
+        }
+    }
+
+    // Normally feature transitions are applied to a bank that hasn't been
+    // frozen yet.  Freeze the bank early to ensure that no account changes
+    // are made.
+    bank.freeze();
+
+    // Simulate crossing an epoch boundary for a new bank
+    let only_apply_transitions_for_new_features = true;
+    bank.apply_builtin_program_feature_transitions(
+        only_apply_transitions_for_new_features,
+        &AHashSet::new(),
+    );
+}
+
+#[test]
+fn test_startup_from_snapshot_after_precompile_transition() {
+    let (genesis_config, _mint_keypair) = create_genesis_config(100_000);
+
+    let mut bank = Bank::new_for_tests(&genesis_config);
+    bank.feature_set = Arc::new(FeatureSet::all_enabled());
+    bank.finish_init(&genesis_config, None, false);
+
+    // Overwrite precompile accounts to simulate a cluster which already added precompiles.
+    for precompile in get_precompiles() {
+        bank.store_account(&precompile.program_id, &AccountSharedData::default());
+        bank.add_precompiled_account(&precompile.program_id);
+    }
+
+    bank.freeze();
+
+    // Simulate starting up from snapshot finishing the initialization for a frozen bank
+    bank.finish_init(&genesis_config, None, false);
 }
